+++ CHANGELOG.md
@@ -2,18 +2,23 @@
 
 All notable changes to `laravel-verification-code` will be documented in this file
 
+## 1.2.2 - 2022-05-06
+
+- Added: parameter to verify method to prevent deleting of the code after verification
+- Added: command to prune old codes from the database
+
 ## 1.2.1 - 2022-03-01
 
  - Fix composer.json file to ensure package can still be used in Laravel 7/8
- 
+
 ## 1.2.0 - 2022-03-01
 
  - Add PHP 8.1 and Laravel 9 support ([#20](https://github.com/nextapps-be/laravel-verification-code/pull/20))
- 
+
 ## 1.1.0 - 2021-02-15
 
  - Add PHP 8.0 support ([#17](https://github.com/nextapps-be/laravel-verification-code/pull/17))
- 
+
 ## 1.0.0 - 2021-01-15
 
  - No changes